Added unittest for timezone_map_util

+++ ui/delegates/timezone_map_util.cpp
@@ -38,7 +38,7 @@ double ConvertLongitudeToX(double longitude) {
   return ((180.0 + longitude) / 360.0 + xdeg_offset / 180.0);
 }
 
-ZoneInfoList GetNearestZones(const ZoneInfoList& total_zones, int threshold,
+ZoneInfoList GetNearestZones(const ZoneInfoList& total_zones, double threshold,
                              int x, int y, int map_width, int map_height) {
   ZoneInfoList zones;
   double minimum_distance = map_width * map_width + map_height * map_height;
