+++ ui/delegates/timezone_map_util.h
@@ -15,7 +15,7 @@ double ConvertLongitudeToX(double longitude);
 
 // Get a list of zone info whose distance to (x, y) is less than |threshold|
 // in a world map with size (map_width, map_height).
-ZoneInfoList GetNearestZones(const ZoneInfoList& total_zones, int threshold,
+ZoneInfoList GetNearestZones(const ZoneInfoList& total_zones, double threshold,
                              int x, int y, int map_width, int map_height);
 
 }  // namespace installer
